Match confirm pw to other styles + improve js
Styles for UX validation of the confirm password field matches the other fields.
Made validation colors variables in CSS and added them to appropriate styles rules.
Combined styles rules where appropriate.
Made the comparePasswordFields() function for use in multiple event listeners.
This commit contains a slight bug where when confirm-pw field is focused its background is white and the other fields do not behave this way.
I prefer the look of this 'bug' and want to incorporate it into the other fields in a future commit.