Finalized form validation styling
Adjusted CSS so that when an input is focused, its background remains white, and the border color and pseudo element indicator
tell the user if the input is valid.  This is acheived with layering of :not selectors and the :valid :invalid and :focus pseudo classes, as well
as the custom classes for the confirm pw field. Added a check in the confirm pw JS so that the input doesn't change in style until
the user has begun to input a value. Changed input labels to be more UX friendly by removing text-transform all caps rule. Likely need to spend more
time in the CSS as it can likely be simplified.